requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
boto3==1.20.33
botocore==1.23.33
distlib==0.3.3
Django==3.2.8
django-filter==21.1
django-storages==1.12.3
filelock==3.3.2
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.4.0
platformdirs==2.4.0
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
virtualenv==20.9.0
whitenoise==5.3.0
